requirements.txt
streamlit
plotly
torch
numpy
pandas
yfinance
ta
sentence-transformers
scikit-learn
faiss-cpu
transformers
accelerate
fastapi
uvicorn
python-multipart
pydantic
feedparser
requests
